+++ devel.R
@@ -1,2 +1,2 @@
 devtools::load_all()
-get_excerpts(remove_dot_git=TRUE, force = TRUE)
+get_excerpts(remove_dot_git = TRUE, force = TRUE)